Fix harmless compiler warning seen with MSVC.
FossilOrigin-Name: e9508ab1f86ac70a3ca5622ce44141bc05186b199bf32fb8739aef36b95aa05c

+++ ext/session/sqlite3session.c
@@ -5189,7 +5189,7 @@ static int sessionRebase(
       sessionAppendByte(&sOut, pIter->bPatchset ? 'P' : 'T', &rc);
       sessionAppendVarint(&sOut, pIter->nCol, &rc);
       sessionAppendBlob(&sOut, pIter->abPK, pIter->nCol, &rc);
-      sessionAppendBlob(&sOut, (u8*)pIter->zTab, strlen(pIter->zTab)+1, &rc);
+      sessionAppendBlob(&sOut,(u8*)pIter->zTab,(int)strlen(pIter->zTab)+1,&rc);
     }
 
     if( pTab && rc==SQLITE_OK ){
